requirements.txt
asgiref==3.10.0
channels==4.3.2
channels_redis==4.3.0
decouple==0.0.7
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
msgpack==1.1.2

PyJWT==2.10.1
python-decouple==3.8
redis==7.1.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
